Analyze/Ja/mecabif.pm: streamline character normalization.
In 'canonify' subroutine, we used to have our own character
normalization logic implemented as a sequence of s///
substitutions.  This (1) looked unoptimized, (2) felt ugly and
(3) was an extra piece of code to maintain.

This patch revamps the way the characters in Japanese are
normalized.  We borrow the conversion table from standard
Encode::JP, and generate necessary substitution regexp and table
upon start-up using that data.  It also adds a few helper
subroutines to allow anybody to later add customized
normalization rules if they so desire.

After benchmarking, using a few hundred newspaper articles as a
sample input (4.5k lines), the performance worry turns out to be
unfounded.  The difference from the previous version is
unmeasuablely lost in the noise (this most likely is because the
implementation of s/// operator is highly optimized).

Even with the added comments and a few subroutines for easier
customization, the patch removes more lines than it adds, and
corrects readability and maintainability issues.  Incidentally,
by using a more complete set of tables, it normalizes more
characters (most notably, symbols such as '#' spelled in
full-width Japanese) than before.

Junio C Hamano


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@3805 593ce9bc-2028-0410-bb98-c566b643a948