Removed /help handler and replaced with a real workspace.
After some discussion with Ken we decided to remove the /help handler (which
redirected to help-XX where XX is your locale).  Instead /help will always be
English help and then help-XX will be the help for locale XX.  To avoid
confusion help-en will not exist.

This required a small change to the UI where the "Help" link is changed from
pointing at /help to pointing at the right workspace for your locale (i.e.
help for EN or /help-XX for non-EN).

This means we can remove a bunch of complicated code and just rely on actual
workspaces, and a much smaller code-change.

Along the way I also added @COPYRIGHT@ to a few files, and I updated the
help-ja tarball to be current.  I also removed the dubugging code in the
KinoSearch analyzers since it was annoying me.

I am not sure if all tests are passing, they are running now.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@3819 593ce9bc-2028-0410-bb98-c566b643a948