Update mecab dictionary generation procedure.
It helps to tweak the initial word separation rule defined in char.def
file in the source, in order to handle complex words that consists
solely of Katakana characters, that are not in MeCab's dictionary.

The default rule says "if a sequence of katakana characters is seen,
take the maximum run, and if that does not appear in the dictionary,
give it as a whole word without any attempt to break it down".  This
patch updates the rule to say "in such a case, see if the words can be
split into two or more words, if earlier parts of the sequence is a
dictionary word, take it as a freestanding word and process the rest
the same way".  Because foreign words borrowed into Japanese are often
spelled without word boundary space in the original language, this
sometimes helps indexing such words.

[NOTE: I cleaned up Junio's patch to the conversion script a bit, no
functional changes just some code tweaks. -matthew]

Junio Hamano


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@3848 593ce9bc-2028-0410-bb98-c566b643a948