Do not use separate userdic.
Text::MeCab seems to be seriously broken in that it does not seem to
grok a separate userdic (although the command line interface works fine).
The primary reason we would want to use userdic is so that we can keep
customized changes to the dictionary separate from the upstream dictionary,
but because we regenerate the system dictionary from the source anyway,
we can achieve the separation at the dictionary source level.  Using
a physically separate user.dic would be just a "nice-to-have", not
absolute necessity.

This needs to be accompanied with a change to the dictionary conversion
script, so that it leaves the copy of dictionary sources in place.  Then
the user dictionary source can be placed in that directory beforehand to
be included in the resulting "system" dictionary.

Junio C Hamano


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@3986 593ce9bc-2028-0410-bb98-c566b643a948