Add a sample user dictionary and prepare a test case for it.
This adds a sample data for user dictionary and adds a test
case to make sure it is picked up.  The word in the added
user dictionary is not found in the default system dictionary,
so finding the word would be a good indication that the user
dictionary is working as expected.

Junio Hamano


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@3988 593ce9bc-2028-0410-bb98-c566b643a948