mecabif.pm: Update preprocessing strategy.
Earlier we squashed whitespaces because a single Japanese word
often straddle two lines across the EOLN, but a side effect of
this is that two Japanese words spelled with an explicit whitespace
in the middle of a single line were concatenated into one before
given to MeCab.  This corrects the mistake and removes only EOLN,
allowing such a whitespace to explicitly delimit words inside
Japanese strings.

Another preprocessing mistake was that Half-width to Full-width
and other normalizations were attempted in a single regexp replace.
This does not work, if we need to normalize Half-width to Full-width
and then rescan the result to further normalize small-A to normal-A.
Split the normalization regexp into three steps and run them in
sequence.

Adds a new test pattern to t/test-data/ja/mecabif.in (and ja-test.data).

Search-ja.t contained a test that expected an punctuation mark to be
indexed, which would never happen.  Remove it.

Junio C Hamano


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/i18n-p4@4013 593ce9bc-2028-0410-bb98-c566b643a948