Re-wrote the pre-check to query the catalogs.
I was having real problems getting the "describe missing_table" tactic to work reliably.  It 
appears that if backtics are used to store the output from that command into a string, but the
table doesn't exist that the psql exits with a 1 and that exit value gets propagated out of 
this script.

Anyway, using the backtick with describe wasn't working.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/fix-inter-wiki-search-migration@5148 593ce9bc-2028-0410-bb98-c566b643a948