I'm making Test::Live kill all apache daemons it started in an END block.  This is to make it so live tests don't leave apaches running and holding the database open.  This is what has been causing folks that try to run regular tests after live tests to get errors about their testing database being left open.
I also noticed that ApacheDaemon::wait_for_startup was only waiting 3.1 seconds for the apache daemons
to start.  On my dev-hardware with nothing else running, it usually takes about 2 seconds for the 
pid file to get created.  I think this short timeout is at the root of all the timeout failures
we see on teststaging.  So, I doubled the number of attempts in this polling loop and added the amount
of time we waited in the error message.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/aaa-candidate-108@6008 593ce9bc-2028-0410-bb98-c566b643a948