Fix the bug where the drop-down menu is shown above the lightbox. This is a known IE6 bug that can be solved by putting an iframe behind the background shade.
Widget.Lightbox.prototype.show is overriden -- only on IE 6 -- and old
hacks to toggle the drop-down menu on-the-fly are removed. Those will
certainly break when any other element on UI triggers the lightbox again.
So it makes better sense to apply the fix in Widget.Lightbox prototype.

The fix is derived from this jquery plugin:

    http://blog.brandonaaron.net/my-jquery-plugins/background-iframe/


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/bug-413-for-spikesource@6342 593ce9bc-2028-0410-bb98-c566b643a948