Downgrading the CGI dependency back to 3.10 since that's the latest version that works for our code on our production systems.
There appears to be some issue with redirect handling and rewriting when
using 3.29 and 'regular, unspecified' ports for http and https and CGI
rewriting them to include ports.  Which for the dual apache setup ends
up being the wrong thing.

Anyway, JJP's plan is to merge this to trunk when it passes QA and immediately
upgrading everyone's open branches with the new trunk code so they all get the
downgrade.

After that, someone should go and make sure talc, topaz, galena, etc all have
the same CGI version.

After that, someone should figure out a better process for vetting dependency 
upgrades that really tests all wacktastic aspects of ST code so keeping current
is less painful and opaque.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/aaa-candidate-112@6481 593ce9bc-2028-0410-bb98-c566b643a948