Set static location for IE 6 behavior CSS
The URL for IE 6 CSS behavior files is relative to the page not the CSS
file that loads it (even though the spec says relative to the CSS file). The 
the .htc file needs to live outside the skins folders so a simple relative path
can be used regardless of what skin the workspace is using. I've put the htc
file at this level and will then remove the .htc file from the S2 skin's
javascript folder.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-22feb@6767 593ce9bc-2028-0410-bb98-c566b643a948