revert the removal of handling old attachment URLs
Our app should only be outputting REST uris for attachments,
so old style attachment URLs *should* not be necessary anymore.

But I'm choosing to leave this handler in for a while, so old URLs
still work.  In the future, we'll remove this handler and the old 
attachment handling code too.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-22feb@6778 593ce9bc-2028-0410-bb98-c566b643a948