Removed call to reports setup code.
The reports setup doesn't work if it is sudoed to www-data, which you need to
do in order to call create-test-data-workspace on a ve.

If you want to test reports, just run st-populate-reports-db as yourself from
the command line.  That will create and import the data you need and also give
devnull1 permissions on /nlw/control


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-04-04@7088 593ce9bc-2028-0410-bb98-c566b643a948