We want to generate summaries regardless of whether we request them.
By default, we request titles only in most of our listviews, which
builds a cached result set _without_ summary data. When we then request
summaries, we read that cached data and no summaries are returned.

If we force the summaries to be read in regardless, they will be available
when we read the cache, therefore we can still reap the benefits of our cached
data. Yay!

brandon and stash


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-06-13@8042 593ce9bc-2028-0410-bb98-c566b643a948