Update Socialtext::Schema so that we can specify a from and to schema version.
We're doing this because migrations run before schema updates, and it may be the
case that a migration needs the schema to exist at a version that has not been
loaded yet.

We would run in to problems if we were do just update the schema to the current
versions, a la migrations 8, 9, and 10 in the past, where migrations were running
against a database from the future, whoa!

-blando


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-07-25@8722 593ce9bc-2028-0410-bb98-c566b643a948