Fix a bug in the workspace re-index.
John T. noticed on prod that workspace re-indexing had ground to a halt. We
dug in and noticed that the re-indexer was failing silently ( and causing
the rampup process in general to stall indefinitely ) when the workspace that
was being indexed did not exist. We added a simple check to ensure that
we index only exisiting workspaces.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-08-08@8964 593ce9bc-2028-0410-bb98-c566b643a948