Update gen-config to fix {bz: 854}. There is a problem because upgrades from really old, pre versioned search code were getting the new ( version 2 ) search functionality without a call to create version 2 indexes, so they were searching an empty index list and returning no results.
This will use the new functionality in st-process-rampup-index, now
renamed to just st-process-index in order to create the new indexes.

While this re-index is happening, searches will only return results for
data that has been indexed. The same issue occured when we switched to version
1 searches and just because some machines are skipping that index does not
mean that they don't have to pay that price.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-08-08@8970 593ce9bc-2028-0410-bb98-c566b643a948