Fix {bz: 1245} by escaping prohibited characters in LDAP filter (search) terms.
The fix covers general searching as well as authentication.  With auth, the
bug occurs when we search for a user in anticipation of referrals.

This may fix things for non-ASCII usernames as well.

The fix should be manually tested against a real ActiveDirectory server to
check that it can handle escape characters in filter statements.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-09-19@10188 593ce9bc-2028-0410-bb98-c566b643a948