Fix {bz: 1346}
The path to the edit icon is coded by the wafl generator and not in a template. S2 and
S3 store images differently: S2 uses multiple directories, S3 does not. To support both
formats, I updated the code to check to see if the s3 skin is specified and path
accordingly. This works for S3 but I am not sure how to do this for new custom skins
that derive off S3.
 

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-09-19@10205 593ce9bc-2028-0410-bb98-c566b643a948