Fixes the blocker {bz: 1416}.
What's happening ?

The resizeEditor() method is adjusting the toolbar width according to the
edit pane, which is either an iframe or a textarea, and has a chance to
have real small width (0 or 1px). The toolbar is then set to be 2px wide.
Therefore all the icons reflows into one vertical line.

This fix add a guardian if-statement to test if the width of edit pane
make sense.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-09-19@10308 593ce9bc-2028-0410-bb98-c566b643a948