This migration is failing as of the iteration-2008-10-17 branch because of the overhaul done to the ST::User* classes. When this migration runs, the new version of these classes is in place. However, the underlying database structure has _not_ been updated. This results in lots of badness as ST::User queries against tables that don't exist.
In order to solve this problem and save ourselves from re-writing this migration from scratch, we're adding an extra 'use' directive into the migration which will use the lib directory found in the same dir as the migration executable. Within this lib is a copy of all of the old fully functional ST::User* modules.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-10-17@10707 593ce9bc-2028-0410-bb98-c566b643a948