Unit test to bang on the short/long term user caches and make sure that they work as expected.
Includes some basic perf testing to demonstration that yes, this stuff really
*does* make a difference.  Run with NLW_BENCHMARK=1 in your environment to
enable the perf testing.

Perf testing is based on a simple "napkin math" scenario where we fire off a
bunch of User lookups that'd be equivalent to what we'd be doing in a regular
HTTP request.  At the end of each request, the short-term user cache is
flushed.

One of the variables in the perf testing is the "delay" that is imposed by the
LDAP lookup.  This was done to help make it more obvious to the reader that
depending on the amount of time that it actually takes to go out and do an
LDAP lookup, the performance improvement that we can see from having these
caches can vary widely.

Regardless, though, having the caches enabled makes a *HUGE* difference.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-10-17@10726 593ce9bc-2028-0410-bb98-c566b643a948