Don't try to cache-bust the long-term LDAP cache, just turn it off during the unit test.
An instrumentation hook was added back in r10715 which can be used to turn
on/off the long-term LDAP cache in the DB.  Instead of calling
`$user->expire()` to try to bust past this cache, just turn the cache off.

Issue with trying to cache-bust the long-term LDAP cache is that when you
expire a user it not only expires their DB entry but also removes them from
the short-term cache automatically.  Next time the user object is
instantiated, we go out and refetch them and then add them to the cache.  When
we only had a short-term cache we wanted to avoid that behaviour (as any
update would necessitate another LDAP lookup) but with the long-term cache in
place this now only necessitates a DB lookup (which is far *less* expensive).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-10-17@10744 593ce9bc-2028-0410-bb98-c566b643a948