Fix {bz: 1542} by always ignoring the breadcrumb_result_set cache file.
The file was previously written to when a user first navigated to the
'recently viewed' page and read from when the user changed the sort order.
However, a change made to fix the recent changes behaviour broke this.

Given that we're only dealing with 25 pages, and given that we don't allow the
user to re-sort the pages when viewing, there's no performance need to cache
this result set.  As such, we can safely just use a fresh 'result_set' every
time.

Leave these orphaned 'breadcrumb_result_set' files around until we do a disk
cleanup migration later.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-10-17@10816 593ce9bc-2028-0410-bb98-c566b643a948