Make the core system more fault tolerant. With the new backup in place, it is possible that there will be missing index.txt files for workspace pages; there is also the possibility that the page dir will exist but be empty -- both of which cause app errors.
With this code, we'll check that the index file exists, if not, we'll
ascertain which revision file it should point to. If there are no
revision files, the app will treat things as though the page doesn't
exist.

These are some very core changes very late in the game, but they are
necessary in order to ship our backup/restore code.

** Please take some time to review this code.**


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/branches/iteration-2008-10-17@11075 593ce9bc-2028-0410-bb98-c566b643a948